Add stack frame lookup based on runtime.CallersFrames for supported go versions
Fix loop


Fix build tags
