+++ internal/stacktrace.go
@@ -24,14 +24,6 @@ type stacktraceFrame struct {
 	Line int64
 }
 
-func (st StackTrace) frames() []stacktraceFrame {
-	fs := make([]stacktraceFrame, len(st))
-	for idx, pc := range st {
-		fs[idx] = lookupFrame(pc)
-	}
-	return fs
-}
-
 func (f stacktraceFrame) formattedName() string {
 	if strings.HasPrefix(f.Name, "go.") {
 		// This indicates an anonymous struct. eg.
@@ -62,27 +54,6 @@ func (f stacktraceFrame) WriteJSON(buf *bytes.Buffer) {
 		w.intField("line", f.Line)
 	}
 	buf.WriteByte('}')
-}
-
-func lookupFrame(pc uintptr) stacktraceFrame {
-	// The Golang runtime package documentation says "To look up the file
-	// and line number of the call itself, use pc[i]-1. As an exception to
-	// this rule, if pc[i-1] corresponds to the function runtime.sigpanic,
-	// then pc[i] is the program counter of a faulting instruction and
-	// should be used without any subtraction."
-	//
-	// TODO: Fully understand when this subtraction is necessary.
-	place := pc - 1
-	f := runtime.FuncForPC(place)
-	if nil == f {
-		return stacktraceFrame{}
-	}
-	file, line := f.FileLine(place)
-	return stacktraceFrame{
-		Name: f.Name(),
-		File: file,
-		Line: int64(line),
-	}
 }
 
 func writeFrames(buf *bytes.Buffer, frames []stacktraceFrame) {
